pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-twin-engine"
version = "0.1.0"
description = "Deterministic probing engine to map corporate entities within LLM latent spaces"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Semantic Twin Team" }
]
keywords = ["llm", "semantic", "audit", "embeddings", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    
    # AI Providers
    "mistralai>=0.1.0",
    
    # Math & Vectors
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    
    # Resilience
    "tenacity>=8.2.0",
    
    # Web Framework
    "Flask>=2.3.0",
    "Flask-CORS>=4.0.0",
    "gunicorn>=21.0.0",
    
    # Visualization
    "plotly>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/semantic-twin-engine"
Documentation = "https://github.com/yourusername/semantic-twin-engine/blob/main/README.md"
Repository = "https://github.com/yourusername/semantic-twin-engine.git"

[tool.setuptools]
package-dir = {"" = "semantic_twin_engine"}

[tool.setuptools.packages.find]
where = ["semantic_twin_engine"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = "\\.pyi?$"
extend-exclude = """
/(
    \\.venv
    | venv
    | __pycache__
    | \\.mypy_cache
    | \\.pytest_cache
    | build
    | dist
    | output
    | cache
)/
"""

[tool.ruff]
line-length = 88
target-version = "py310"
extend-exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".mypy_cache",
    ".pytest_cache",
    "build",
    "dist",
    "output",
    "cache",
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=semantic_twin_engine --cov-report=term-missing"
